:recycle:DDDに沿ってリファクタリング（改善の余地あり）
