+++ main.py
@@ -1,18 +1,27 @@
 import sys
+
+
 sys.path.append("src")
 
 from dotenv import load_dotenv
 
-from application.scheduler import add_job, start_scheduler
-from domain.tweet import post_tweet
+from application.usecase.post_camp_gear_sale_use_case import PostCampGearSaleUseCase
+from infrastructure.repository.product_repository import ProductRepository
+from infrastructure.service.product_service import ProductService
+from infrastructure.service.post_service import PostService
+
+
+load_dotenv()
 
 
 def main():
-    load_dotenv()
+    product_repository = ProductRepository(ProductService())
+    post_service = PostService()
 
-    add_job(post_tweet)
-
-    start_scheduler(post_tweet)
+    post_camp_gear_sale_use_case = PostCampGearSaleUseCase(
+        product_repository, post_service
+    )
+    post_camp_gear_sale_use_case.handle()
 
 
 if __name__ == "__main__":
